m9e0

+++ c++_modules/module09/ex00/classes/BtcFile.hpp
@@ -12,5 +12,5 @@ class BtcFile{
         std::ifstream internalDataBase;
         std::ifstream externalDataBase;
         std::map<std::string, float> db;
-        std::map<std::string, int> db2;
+        std::map<std::string, float> db2;
 };
